feat: add search by Id, update and delete ElementGroup

+++ src/main/java/org/example/repositories/ElementGroupRepository.java
@@ -3,5 +3,5 @@ package org.example.repositories;
 import org.example.entities.ElementGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ElementGroupRepository extends JpaRepository<ElementGroup, Integer> {
+public interface ElementGroupRepository extends JpaRepository<ElementGroup, Long> {
 }
